Allow bwf2csv -o option to append to existing file

+++ autoBWF/bwf2csv.py
@@ -18,21 +18,22 @@ def main():
                      "performer", "topics", "names", "events", "places", "owner", "ICOP"]
 
     if args.outfile is not None:
-        output = csv.DictWriter(open(args.outfile, 'w'), output_fields)
+        if os.path.isfile(args.outfile):
+            # if file exists, assume that it is a bwf2csv output file, append rows without generating header
+            output = csv.DictWriter(open(args.outfile, 'a'), output_fields)
+        else:
+            # otherwise, create new file and generate header
+            output = csv.DictWriter(open(args.outfile, 'w'), output_fields)
+            output.writeheader()
     else:
         output = csv.DictWriter(sys.stdout, output_fields)
-
-    output.writeheader()
-    rows = []
+        output.writeheader()
 
     for infile in args.infile:
         metadata = get_bwf_core(True, infile)
         metadata.update(get_bwf_tech(True, infile))
         metadata.update(get_xmp(infile, ["bwfmetaedit", "--specialchars", "--accept-nopadding"]))
-        rows.append({k: metadata[k] for k in output_fields})
-
-    for row in rows:
-        output.writerow(row)
+        output.writerow({k: metadata[k] for k in output_fields})
 
 
 if __name__ == '__main__':
